--- after/plugin/gitsigns.lua
@@ -1,1 +1,0 @@
-require('gitsigns').setup()
