--- after/plugin/lazygit.lua
@@ -1,4 +1,0 @@
-vim.keymap.set('n', '<leader>Gg', ':LazyGit<CR>', { desc = "Lazy[G]it" })
-
-vim.g.lazygit_floating_window_use_plenary = 1
-vim.g.lazygit_floating_window_winblend = 1
