--- after/plugin/lsp.lua
@@ -1,86 +1,0 @@
-local lsp_zero = require('lsp-zero')
-
-lsp_zero.on_attach(function(client, bufnr)
-  -- see :help lsp-zero-keybindings
-  -- to learn the available actions
-  lsp_zero.default_keymaps({buffer = bufnr})
-end)
-
-require('mason').setup({})
-require('mason-lspconfig').setup({
-	handlers = {
-		lsp_zero.default_setup,
-
-		--Add any special LSP configs here
-		powershell_es = function()
-			require('lspconfig').powershell_es.setup({
-				settings = { powershell = { codeFormatting = { Preset = 'OTBS' } } },
-			})
-		end,
-	},
-})
-
-local cmp = require('cmp')
-local cmp_action = require('lsp-zero').cmp_action()
-
-cmp.setup({
-    mapping = {
-        -- `Enter` key to confirm completion
-        --['<CR>'] = cmp.mapping.confirm({select = false}),
-
-        -- Ctrl+Space to trigger completion menu
-        ['<C-Space>'] = cmp.mapping.complete(),
-        -- Navigate between snippet placeholder
-        ['<C-f>'] = cmp_action.luasnip_jump_forward(),
-        ['<C-b>'] = cmp_action.luasnip_jump_backward(),
-        -- remap tab and shift tab because copilot should use those
-        ['<Tab>'] = cmp.mapping.confirm({select = false}),
-	['<S-Tab>'] = nil,
-        ['<C-j>'] = cmp.mapping.select_next_item({behavior = cmp.SelectBehavior.Select}),
-        ['<C-k>'] = cmp.mapping.select_prev_item({behavior = cmp.SelectBehavior.Select})
-    }
-})
-
-vim.api.nvim_create_autocmd('LspAttach', {
-	group = vim.api.nvim_create_augroup('kickstart-lsp-attach', { clear = true }),
-	callback = function(event)
-
-	local map = function(keys, func, desc)
-		vim.keymap.set('n', keys, func, { buffer = event.buf, desc = 'LSP: ' .. desc })
-	end
-
-	--Search for symbols within current document
-	map('<leader>ds', require('telescope.builtin').lsp_document_symbols, '[D]ocument [S]ymbols')
-
-	--Find references for the word under your cursor
-	map('gr', require('telescope.builtin').lsp_references, '[G]oto [R]eferences')
-
-	-- Jump to the implementation of the word under your cursor.
-	--  Useful when your language has ways of declaring types without an actual implementation.
-	map('gI', require('telescope.builtin').lsp_implementations, '[G]oto [I]mplementation')
-
-	-- Fuzzy find all the symbols in your current workspace
-	--  Similar to document symbols, except searches over your whole project.
-	map('<leader>ws', require('telescope.builtin').lsp_dynamic_workspace_symbols, '[W]orkspace [S]ymbols')
-
-
-	-- The following two autocommands are used to highlight references of the
-	-- word under your cursor when your cursor rests there for a little while.
-	--    See `:help CursorHold` for information about when this is executed
-	--
-	-- When you move your cursor, the highlights will be cleared (the second autocommand).
-	local client = vim.lsp.get_client_by_id(event.data.client_id)
-	if client and client.server_capabilities.documentHighlightProvider then
-		vim.api.nvim_create_autocmd({ 'CursorHold', 'CursorHoldI' }, {
-			buffer = event.buf,
-			callback = vim.lsp.buf.document_highlight,
-		})
-
-		vim.api.nvim_create_autocmd({ 'CursorMoved', 'CursorMovedI'}, {
-			buffer = event.buf,
-			callback = vim.lsp.buf.clear_references,
-		})
-	end
-
-end,
-})
