--- after/plugin/toggleterm.lua
@@ -1,23 +1,0 @@
-local powershell_options = {
-  shell = vim.fn.executable "pwsh" == 1 and "pwsh" or "powershell",
-  shellcmdflag = "-NoLogo -NoProfile -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.Encoding]::UTF8;",
-  shellredir = "-RedirectStandardOutput %s -NoNewWindow -Wait",
-  shellpipe = "2>&1 | Out-File -Encoding UTF8 %s; exit $LastExitCode",
-  shellquote = "",
-  shellxquote = "",
-}
-
-for option, value in pairs(powershell_options) do
-  vim.opt[option] = value
-end
-
-require("toggleterm").setup{
-	hide_numbers = true,
-	autochdir = true,
-	start_in_insert = true,
-	auto_scroll = true,
-	size = 30
-}
-
-vim.keymap.set('n','<leader>t',':Telescope toggleterm_manager<CR>', { desc = "[T]oggleterm Manager"})
-vim.keymap.set('v','<leader>r',':ToggleTermSendVisualLines<CR>', { desc = "[R]un selection" })
