--- after/plugin/tokyonight.lua
@@ -1,12 +1,0 @@
-require("tokyonight").setup ({
-	style = "storm",
-	transparent = false,
-	terminal_colors = true,
-	styles = {
-		comments = { italic = true },
-		keywords = { italic = true },
-		sidebars = "dark",
-		floats = "dark",
-	},
-	dim_inactive = false,
-})
