--- after/plugin/treesitter_parsers.lua
@@ -1,9 +1,0 @@
-local parser_config = require("nvim-treesitter.parsers").get_parser_configs()
-parser_config.powershell = {
-  install_info = {
-    url = "https://github.com/jrsconfitto/tree-sitter-powershell",
-    files = {"src/parser.c"}
-  },
-  filetype = "ps1",
-  used_by = { "psm1", "psd1", "pssc", "psxml", "cdxml" }
-}
