--- after/plugin/whichkey.lua
@@ -1,13 +1,0 @@
-require("which-key").setup()
-
---Document existing keychains
-require('which-key').register {
-	['<leader>c'] = { name = '[C]ode', _ = 'which_key_ignore' },
-	['<leader>d'] = { name = '[D]ocument', _ = 'which_key_ignore' },
-	['<leader>dd'] = { name = '[D]ocument [D]iff', _ = 'which_key_ignore' },
-	['<leader>r'] = { name = '[R]ename', _ = 'which_key_ignore' },
-	['<leader>s'] = { name = '[S]earch', _ = 'which_key_ignore' },
-	['<leader>w'] = { name = '[W]orkspace', _ = 'which_key_ignore' },
-	['<leader>g'] = { name = '[G]oto', _ = 'which_key_ignore' },
-	['<leader>G'] = { name = '[G]it', _ = 'which_key_ignore' }
-}
