+++ lua/SrBlackVoid/init.lua
@@ -1,3 +1,3 @@
 require("SrBlackVoid.remap")
-require("SrBlackVoid.packer")
 require("SrBlackVoid.set")
+require("SrBlackVoid.lazy")
