+++ lua/SrBlackVoid/plugins/letitsnow.lua
@@ -1,0 +1,5 @@
+return {{
+	"marcussimonsen/let-it-snow.nvim",
+	cmd = "LetItSnow",
+	opts = {},
+}}
