+++ lua/SrBlackVoid/plugins/lualine.lua
@@ -1,0 +1,9 @@
+return {
+	{
+		"nvim-lualine/lualine.nvim",
+		dependencies = {
+			{'nvim-tree/nvim-web-devicons', opt = true },
+		},
+		config = require("SrBlackVoid.themes.lualine-bubbles")
+	}
+}
