+++ lua/SrBlackVoid/plugins/minidiff.lua
@@ -1,0 +1,10 @@
+return {
+	{
+		'echasnovski/mini.diff',
+		lazy = false,
+		keys = {
+			{'<leader>Gd', ':lua MiniDiff.toggle_overlay()<CR>', desc = "[G]it [D]iff" }
+		},
+		opts = {},
+	}
+}
