+++ lua/SrBlackVoid/plugins/oil.lua
@@ -1,0 +1,18 @@
+return {
+	{
+		"stevearc/oil.nvim",
+		dependencies = { "nvim-tree/nvim-web-devicons" },
+		opts = {
+			columns = { "icon" },
+			view_options = {
+				show_hidden = true,
+			},
+			keymaps = {
+				["<C-h>"] = false,
+				["<C-j>"] = false,
+				["<C-k>"] = false,
+				["<C-l>"] = false,
+			}
+		},
+	}
+}
