+++ lua/SrBlackVoid/remap.lua
@@ -1,10 +1,10 @@
 vim.g.mapleader = " "
 
---Open up NetRW
-vim.keymap.set("n", "<leader>f", vim.cmd.Ex, { desc = "[F]ile Explorer"})
+--Open up File Explorer
+vim.keymap.set("n", "<leader>f", "<CMD>Oil<CR>", { desc = "[F]ile Explorer"})
 
---Open up NeoTree
--- vim.keymap.set("n", "<C-n>", vim.cmd.Neotree)
+--Open up plugin manager
+vim.keymap.set("n", "<leader>l", ":Lazy<CR>", { desc = "[L]azy.nvim"})
 
 -- Move between windows easier
 vim.keymap.set('n', '<C-h>', '<C-w><C-h>', { desc = 'Move focus to the left window' })
@@ -13,7 +13,7 @@ vim.keymap.set('n', '<C-j>', '<C-w><C-j>', { desc = 'Move focus to the lower win
 vim.keymap.set('n', '<C-k>', '<C-w><C-k>', { desc = 'Move focus to the upper window' })
 
 --Remap word deletion during insert mode (not currently working)
--- vim.keymap.set('i', '<C-BS>', '<C-W>', { desc = 'Full word deletion' })
+vim.keymap.set('i', '<C-H>', '<C-W>', { desc = 'Full word deletion' })
 
 --Quickly clear highlighting
 vim.keymap.set('n', '<leader>h', vim.cmd.noh, { desc = 'Clear [H]ighlighting'})
