+++ lua/SrBlackVoid/set.lua
@@ -43,3 +43,17 @@ vim.g.terminal_emulator='pwsh'
 
 --Set shell to PowerShell 7
 vim.o.shell = 'pwsh'
+local powershell_options = {
+	shell = vim.fn.executable "pwsh" == 1 and "pwsh" or "powershell",
+	shellcmdflag = "-NoLogo -NoProfile -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.Encoding]::UTF8;",
+	shellredir = "-RedirectStandardOutput %s -NoNewWindow -Wait",
+	shellpipe = "2>&1 | Out-File -Encoding UTF8 %s; exit $LastExitCode",
+	shellquote = "",
+	shellxquote = "",
+}
+
+for option, value in pairs(powershell_options) do
+	vim.opt[option] = value
+end
+
+vim.opt.colorcolumn = '80'
